requirements.txt
tensorflow-cpu==2.6.2
numpy==1.19.5
streamlit==1.3.0
